rename log.h to sdliv_log.h
