+++ source/log.cpp
@@ -1,4 +1,4 @@
-#include <log.h>
+#include <sdliv_log.h>
 
 #include <iostream>
